+++ src/console/src/factory/services/ledger/mod.rs
@@ -1,2 +1,3 @@
+pub mod cycles;
 pub mod icp;
 pub(super) mod icrc;
