+++ src/console/src/factory/services/mod.rs
@@ -1,3 +1,2 @@
-mod cycles_ledger;
 pub mod ledger;
 pub mod payment;
