+++ src/console/src/factory/services/payment.rs
@@ -1,5 +1,5 @@
-use crate::factory::services::cycles_ledger::transfer_from_cycles_ledger;
-use crate::factory::services::ledger::{transfer_from, verify_payment};
+use crate::factory::services::ledger::cycles::cycles_transfer_from;
+use crate::factory::services::ledger::icp::{icp_transfer_from, icp_verify_payment};
 use crate::store::stable::is_known_payment;
 use candid::Principal;
 use ic_ledger_types::{BlockIndex, Tokens};
@@ -15,9 +15,9 @@ pub async fn process_payment_icp(
             return Err("Payment has been or is being processed.".to_string());
         }
 
-        verify_payment(&purchaser, &block_index, fee).await?
+        icp_verify_payment(&purchaser, &block_index, fee).await?
     } else {
-        transfer_from(&purchaser, &fee).await?
+        icp_transfer_from(&purchaser, &fee).await?
     };
 
     let ledger_id = Principal::from_text(ICP_LEDGER).unwrap();
@@ -30,7 +30,7 @@ pub async fn process_payment_cycles(
     _block_index: Option<BlockIndex>,
     fee: Tokens,
 ) -> Result<(Principal, BlockIndex), String> {
-    let purchaser_payment_block_index = transfer_from_cycles_ledger(&purchaser, &fee).await?;
+    let purchaser_payment_block_index = cycles_transfer_from(&purchaser, &fee).await?;
 
     let ledger_id = Principal::from_text(CYCLES_LEDGER).unwrap();
